+++ r/NEWS.md
@@ -1,5 +1,8 @@
 # mooplot 0.1.1
 
+ * Requires `moocore >= 0.1.5` because `moocore` renamed `vorobT()` and
+   `vorobDev()` to `moocore::vorob_t()` and `moocore::vorob_dev()` to be consistent with other
+   function names.
  * Fix bug in `eafdiffplot` where colors were plotted in the wrong order.
 
 # mooplot 0.1.0